feat: add login and logout steps to post page tests
Implemented login and logout steps in the post page tests to enhance user interaction coverage. This includes visibility checks for sign-up and sign-out dialogs, ensuring proper functionality during user sessions.
